Move ErrorList into errors package
