+++ internal/errors/errors.go
@@ -1,4 +1,4 @@
-package parser
+package errors
 
 import (
 	"fmt"
