+++ internal/ttcn3/parser/parser.go
@@ -5,6 +5,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/nokia/ntt/internal/errors"
 	"github.com/nokia/ntt/internal/loc"
 	"github.com/nokia/ntt/internal/ttcn3/ast"
 	"github.com/nokia/ntt/internal/ttcn3/scanner"
@@ -14,7 +15,7 @@ import (
 // The parser structure holds the parser's internal state.
 type parser struct {
 	file    *loc.File
-	errors  ErrorList
+	errors  errors.ErrorList
 	scanner scanner.Scanner
 
 	// Tracing/debugging
